Update examples of Drush commands
Both examples use php-eval.  The first example omits the trailing `;`, which Drush will supply automatically.  This will pass through PantheonSSH and print 'oh happy days'.

The second example has an evil, deprecated `;` character in it, which PantheonSSH despises, and will reject.  The result will be 'Command not supported as typed'.

As a follow-on task, it would be nifty to consider loosening up PantheonSSH to allow more flexible command processing.  The current set of rejected characters are designed to prevent the user from accidentally splitting a single command into two commands; however, it is currently possible to do this via `&&`, so perhaps relaxing the restrictions further would not have a very large impact.